+++ CHANGELOG.md
@@ -9,6 +9,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 
 - Created a loader method for git notes to support things like, infinite scrolling in webview, caching etc (<https://github.com/jrosco/vscode-git-notes/pull/19>)
+- Remove and Add committed notes from the `CommitDetails[]` interface class (<https://github.com/jrosco/vscode-git-notes/pull/18>)
+  
+### Fixed
+
+- Don't show files in webview if no file changes found (<https://github.com/jrosco/vscode-git-notes/pull/17>)
 
 ## [0.1.0]